mac: Add a direct conversion from Mac<F2, F40b> to Mac<F40b, F40b>.
Admittedly, I'm not 100% sure this is a legitimate/safe thing to do -
but given what I do understand about how these values are generated
and communicated, I don't believe there should be any difference
between computing with F2 values as F2 values and as F40b values,
given the subfield relationship between the two.

There is a TODO in the code questioning safety, for my own
sanity. That said: This commit implements From to convert Mac<F2,
F40b> values to Mac<F40b, F40b> values in the natural/obvious
way. This will be used to lift the input wire values to F40b in the
implementation of disjunctions for binary circuits.

Additionally, this commit adds a TryFrom instance in the opposite
conversion direction. This is used after evaluating the disjunction to
lower the output wires back to F2.

These operations should be safe since F2 is a subfield of F40b; that
means that any operations between F2 values have exactly the same
semantics in F40b.