backend_extfield: Implement helper function to lift the guard.
To implement disjunctions in F2, we lift the guard value (which is
given as one or more condition wires in big-endian order) to a single
F40b value. We do this to Macs, in particular, using the lift() method
and zero-padding with empty/zero Macs as necessary.

This helper is local to the disjunction method, and assumes that its
parameter `inputs` matches the parameter specification of the
disjunction interface.