backend_extfield: Implement adjust_inputs helper function.
Since we are computing the disjunction over F40b, we must re-interpret
the input Macs (which are over F2) as Macs over F40b. This is the
use-case for the earlier-implemented From/TryFrom instances on Boolean
macs.

Note also that the Vec returned by this function replaces the original
condition wires with the guard Mac that is computed by the helper
lift_guard implemented in the previous commit. This will be reflected
in the compilation of Disjunctions, where an earlier commit permitted
the specification of the number of condition wires to allow for this
conversion ahead of the running of the Dora protocol.