backend_extfield: Implement execute_branch helper function.
This commit implements the "meat and potatoes" of binary
disjunctions. In particular, it combines the lift_guard and
adjust_inputs helper functions to ultimately run dora.mux to execute a
branch of the disjunction.

We also see here the TryFrom instance between F40b and F2 Macs. This
should be infallible, as the original wires were all F2, and F2 is a
subfield of F40b.