test_circuits: Add a simple F2 disjunction test.
To test out the F2 disjunction support, this commit adapts the example
in the disjunction specification and includes it with the other test
circuits.

This circuit has been tested and appears to run as expected. Both
branches have been tested for both success and failure under the given
input bits.