dora/translate: Use a BTreeMap for the SIEVE WireId <-> Cell map.
Previously, Dora utilized a Vec of Option<WireId> to map SIEVE IR wire
IDs to R1CS cells.

This choice of data structure is problematic, however, since the
indices can be very large (i.e. near u64::MAX), thus requiring
allocations of that size to use a Vec as a sparse array.

This commit replaces this problematic use of Vec with a BTreeMap to
save memory without hurting performance too badly. As a bonus, this
change also simplifies the code through use of the Entry API on
BTreeMap (and, of course, no longer dealing with the Options as the
value type).