ram/mod: Implement a BinaryCounter type.
To support RAMs over F2, we need to implement a binary counter over
the address size number of bits.

This commit adds such a structure, with some unit tests for the
incr method that does all the interesting work.

Note that currently, the curr_val method allocates a new Vec to
represent the current value. These allocations will be generally
small and short-lived, only being created during the finalization of a
RAM when all addresses are enumerated.