ram/tests: Add a test for Boolean RAM.
This test is similar to test_arithmetic_ram, generating RAM_STEPS
random addresses and 'round-tripping' with the initial value (that is,
performing a remove and insert, which is like doing a read/write).

For the addresses, we use 7 bits, where the leading bit is always
0 (so we don't have to worry about checking whether the address is <
RAM_SIZE). Values are similar represented in 7 bits.