ram/mod: Use compile-time constants for challenge sizes.
Previously, we directly passed an integer value to DoraRam::new for
the challenge size of both arithmetic and Boolean RAMs.

This commit introduces compile-time constants holding these values, so
they can easily be updated if necessary. Furthermore, the value for
Boolean RAMs is increased to 2^40 (ie 1 << 40) to reduce the soundness
error to negligible/appropriate for our security parameter of 40.