mod.rs: Remove the PluginBody type.
Previously, the PluginBody type applied #[allow(dead_code)] to two of
the three structure fields, corresponding to the plugin name and
operation.

These fields (name/operation) are never used, and are indeed
redundant (as the plugin name is captured in the Plugin trait, and the
operation has already been ingested and compiled to a
PluginExecution). Effectively, this makes PluginBody an unnecessary
newtype wrapper around PluginExecution.

This commit removes this layer of indirection by removing the
PluginBody type, using PluginExecution directly instead.