Clean up spurious allow() directives.
There were a number of places where we used allow() to suppress
warnings either redundantly (ie, in places already allowed by a
higher-level allow()) or in place of cleaning up imports.

This commit corrects these uses of allow(), including cleaning up the
imports when possible. Note that in some cases, this involved the
deletion of dead/unused code.