lib: Make the edabits module private.
Edabits is a subprotocol of Diet Mac'n'Cheese used to implement field
conversion. While it is possible that users of swanky will want to use
this protocol in isolation / in other projects, if that is the case,
the protocol should be implemented in a separate crate from Diet
Mac'n'Cheese.

As the implementation was not done with this use in mind (and indeed,
since it depends on other Diet Mac'n'Cheese-specific libraries), this
commit simply makes the module private to reflect its status as an
implementation detail of Diet Mac'n'Cheese.