lib: Make the backend_multifield module private.
The Diet Mac'n'Cheese binary (and, similarly, the example use of Diet
Mac'n'Cheese as a library and Web Mac'n'Cheese) only need access to
the EvaluatorCirc type from backend_multifield.

Types such as DietMacAndCheeseConv are implementation details (see
previous commits for reasoning that applies in this situation as
well).

This commit hides a significant number of internal details from anyone
looking to use Diet Mac'n'Cheese as a library, cleaning up the public
API significantly.