dietmc: Multihtread F2 extension field backend
This is done by changing the constructor for backend containing the extension field voles,
by making it take as argument a Fcom functionality for F40b,F40b voles.
In the backend_multifield, the multithreaded evaluator spawn the F40b,F40b voles
in addition to the F2,F40b voles.
The F40b,F40b voles functionality is added to the evaluator struct.