circuit_ir: Remove the unnecessary duplication of function names.
We previously provided the method id_to_name to convert internal
function IDs back to user-friendly string names.

This method is completely unused in DMC, Web Mac'n'Cheese, and any/all
example files. Furthermore, it is only for the sake of this method
that we stored the function name alongside its declaration in the
internal FunId-indexed Vec.

This commit simplifies the representation by removing this method, and
the extra storage of a function's name.