backend_multifield: Simplify push_elem using the Entry API.
Previously, push_elem was implemented using an awkward combination of
the Entry API (to initialize with a default if necessary), and a chain
of method calls that massage the result into a mutable reference.

These can be combined using the Entry API alone, and in a way that is
cleaner / doesn't rely on so many explicit conversions between
reference types.