dmc: When appropriate, use ? instead of unwrap().
This commit replaces a number of uses of unwrap() with the ?
operator. In particular, this was done in functions returning a Result
where the unwrap() was inappropriately applied to an intermediate
result, or where Option values could return more meaningful errors.

Care was taken to leave unwrap() whenever some invariant was
guaranteed to hold (e.g. on an internal helper function call). In
these cases, there's no reason to expand on the panic.