popsicle/circuit_psi: Create proptest strategy for generating inputs
This will be useful for almost all circuit psi tests. We also have strategies for payloads
as u128 and Block512 because the later helps us avoid manually turning payloads as Vec<u128>
into Vec<Block512> that circuit psi expects. We this strategy when we generally
just want to check that the protocol has suceeded with no errors and do not need
to read or operate on the payloads. We use the `arbitrary_payloads_u128` for the later case.