popsicle/circuit_psi: Handle unused variable
popsicle/circuit_psi: Remove unused import

popsicle/circuit_psi: Make errors module public for examples

We need to do this in order for users to be able to define their own circuits outside of swanky (in which case they would need to restrict the error type).

popsicle/circuit_psi: make ci happy

popsicle/circuit_psi: Fix bug in hashing test

The original test checked that the output of the hashes intersects with the original set. This makes no sense because hashing changes the inputs and maps them to their hashed values. Now we check that the output of both hashes preseves the intersection because this is what ultimately matters.