popsicle/circuit_psi: Switch all circuit psi to unit tests
When running proptests in debug mod, as is the case for CI, the tests are
exceedingly slow and CI is failing. I believe that this may be to things
like the complexity of OPPRF initialization, or the hashing etc. There is
an argument to be made here that proptest is not well suited for these functions
anyway because of how complex they are.