homcom: Generate a new chi on multiplication check reset.
The unsoundness of our QuickSilver implementation related to the
generation and use of the challenge value chi.

This commit generates a new chi challenge each time the multiplication
check state is reset (which occurs each time the current state is
finalized).

Since generation of a fresh chi relies on RNG, we thread an AesRng
through the finalize method.