homcom: Update the accumulate method for multiplication checks.
Previously, the accumulate method modified internal state variables to
hold the prover-private sums.

Since doing this for all checks leads to unsoundness, this commit
updates the method to simply push to the internal queue in the prover
case.

Note that accumulate now needs the mask, channel, and RNG; these will
be used to, when appropriate, finalize a batch of multiplication
checks that we have accumulated in the queue.