homcom: Update the multiplication check finalize method.
To match the previous commit, this moves the value-accumulation logic
for provers to the finalize method. Furthermore, the accumulate method
now finalizes the multiplication check state if the number of checks
meets or exceeds the BATCH_SIZE.

Notice that we must compute the sum_a0 and sum_a1 locally now, as well
as chi_power - this will allow us to simplify the internal state
significantly.