homcom: Make internal chi_power VerifierPrivateCopy.
To avoid confusion between the prover's locally-computed chi_power and
the internally-accumulated chi_power used in the optimized verifier
behavior, this commit makes the chi_power field of the multiplication
check state VerifierPrivateCopy, so it can _only_ be used in verifier
computations.

This isn't strictly necessary, especially since chi is shared and the
number of checks is public, but helps with cognitive separation of the
concepts.