homcom: Make chi VerifierPrivateCopy.
The stored chi value should only be available to the verifier while
accumulating; this commit changes the type of the multiplication check
state's chi field to be VerifierPrivateCopy to reflect this.

This has a number of important downstream consequences. First and most
important, we no longer communicate chi every time it is generated -
including the first time, during initialization. This results in some
simplifications to that method, as it no longer needs a channel.

This also impacted the reset method, which now simply generates the
new verifier-private chi, deferring communication.

All communication is now in the finalize method: It starts by sending
chi from verifier to prover, so the prover may compute its private
sums honestly.

Ultimately, these changes made the associated chi() method obsolete,
as we no longer couple the generation of chi to its communication
between parties - and this communication only happens in one place.