homcom: Make BATCH_SIZE pub.
This commit anticipates a future change, where clients of the
MultCheckState type are responsible for calling finalize on
batches (rather than leaving this responsibility to the accumulate
method).

This is to avoid excess communication when generating masks for the
finalize step; currently, we generate masks for every accumulate call,
but these are only used once every 3000000 calls, wasting a lot of
communication on SVOLEs that are ultimately never used.