homcom: Move batched finalization out of accumulate.
Previously, we ended up doing a large amount of wasted work to
generate mask values that were only actually used once every 3000000
calls to accumulate.

This commit fixes this waste by moving the batched finalization out of
accumulate and into clients of MultCheckState. This means accumulate's
type can be restored to its previous (much simpler) state, and the
only 'cost' is a check on the number of accumulated values after every
call to accumulate in client code.