deny.toml: Remove all deprecated keys.
The most recent version of cargo-deny changed a significant number of
lint levels to be deny by default. Additionally, the notion of
severity-threshold was completely removed.

Of particular note is the relative simplicity with which licenses are
now managed; all previous rules have been replaced with more
aggressive default denial and only the notions of `allow` and
per-crate `exceptions`.