FIXME: Duplicate deny(missing_docs) attribute where necessary.
IMPORTANT! This commit should be reverted as soon as an upgrade to
tree-sitter's Rust grammar makes its way into nixpkgs.

There's some kind of regression / issue in tree-sitter's Rust parser
in our Nix setup that results in the first inner attribute being
parsed as a shebang instead of an attribute.

I don't fully understand this; I found a commit specifically fixing
this issue in the tree-sitter Rust repository that is more than a year
old. Not to mention the fact that we did not have this issue before!

This commit offers a crappy temporary fix: Double-up on the
"#![deny(missing_docs)]" where necessary.