homcom: Move all communication to check-zero finalization.
This commit moves the too-early communication between prover and
verifier to the finalize method to address a potential soundness bug,
and updates downstream modules to match.

This effectively turns the ZeroCheckState type into a strongly-typed
newtype around the queue of wires to check; this was the change with
the least impact on downstream code (at most needing to add a check
analogous to that for QuickSilver in the Edabits implementation, to
finalize early if the batch has filled up).

Note that this also removes the prover-private wrapper around the
queue; unlike QuickSilver, this protocol has symmetric behavior
between prover and verifier (we can't, e.g., accumulate the chi value
early).