popsicle/circuit_psi: Create a struct for the intersection results.
Create a struct for the intersection results, i.e. the set intersection,
the bit vector showing the element existences and the unmasked payloads.
We do so to avoid storing this result in the PsiParty and instead return
this result when `intersect` is called for the user to manipulate these
oblivious values as they please by for example applying a circuit of
their choice to this intersection.

This change is reflected in both the PsiGarbler and the PsiEvaluator that
implement how the intersection happens, along with the affected tests and
examples.