ci: upgrade to newest version of tree-sitter
The preferred way to access a tree-sitter grammar from python is to use
a bespoke python extension which is statically linked against the
compiled grammar. This python extension then gives a handle to the
grammar.

Unfortunately nix doesn't support generating these bindings from the
tree-sitter libraries that it provides. To work around this, we manually
implement the old approach to extract the grammar out of the library
directly, without any C extension.