etc: cache nix-shell environment
Running ./swanky is slow! It takes 13 seconds on my laptop just to get
to the swanky help screen. nix-shell takes up all of this time. To avoid
having to run nix-shell all the time to setup our environment every time
we run ./swanky, this commit uses a python script to cache the nix
environment and then re-use this cache if the nix config files haven't
changed.

After this commit, ./swanky (with a populated cache) takes 0.1-0.5ms.