Refactoring the circuit PSI implementation
This commit makes the CircuitPsi trait simpler by removing the
associated type `F` and the generic type `Party`. The later had to
be specified by the user whenever they called the `intersect` method.

The `PsiGarbler` and `PsiEvaluator` now carry the choice `BasePsi` as a
phantom data type.

The user now can also use a type alias for `PsiGarbler` and `PsiEvaluator`
that alleviates the choice of `BasePsi` all together. This is a major
life improvement.