Turn payloads into Option type and return errors when set sizes !=
This is a lifetime improvement. When calling `intersection_with_payloads`
the payload parameter is now an Option type. This improves the readability
of the code and reduces the possibility of misusing the API that arose from
interpreting an empty slice as the non existence of payloads.

Additionally, this commit checks that the payload set is complete (i.e.
has the same size as the set of elements) and returns an appropriate error
when that is not the case.