vectoreyes: move array conversions to rust
If we were to remove the old conversions entirely, it would break the
vectoreyes test suite (since a completely separate scalar backend gets
generated to test against, but the rust code only generates methods for
the non-test implementation).

To work around this, rather than removing the old implemenation, we only
emit it for the scalar test implementation.