vectoreyes: enable generated vector internal types
We check for force_scalar_for_test in the python template because the
python code generates a scalar instance of the vectoreyes backend for
the test suite (since the rust codegen isn't generating the test suite
yet, and we still want to have tests). So, rather than ripping out this
code, we just disable it for the main backend, and keep it enabled for
testing.