etc: expose rust commands via symlink
By default, the rust-overlay that we use for nix has gcc as a
"propagated build input." This means that anything which depends on
rust-overlay will end up having gcc added as a dependency. This was
overriding our choice to use clang as our C compiler, which was breaking
our cross-compilation support.

To fix this, this commit changes our rust-toolchain derivation to not
_directly_ depend on the rust compiler, but rather on a series of
symlinks to the rust compiler binaries.