etc: Add musl headers nix file
Rust ships with a musl libc.a file, which is sufficient for compiling
just rust code. However, it doesn't ship with C headers which are
neccessary for compiling C code against musl. (And some of our
dependencies include C code.)

This commit adds a nix derivation to fetch and compile the musl headers
for a particular architecture. It also writes a version.txt file which
will be read (in a later commit) to ensure that the version of musl that
rust ships matches the headers that we've downloaded.