ci: add emulator support to caching test runner
This commit adds an environment variable to our caching_test_runner
script which will be used as an interpreter for any test the
caching_test_runner runs.

In a future commit, we'll pass a qemu wrapper script to this script to
let us run test binaries compiled for a different architecture (than the
host architecture).